requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
pyngrok==7.0.1
pytest==7.4.3
pyyaml==6.0.1
pymongo==4.6.0
mongomock==4.1.2

